socialmedia/requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
Django==3.1.11
django-audittrail==0.2.1
django-crispy-forms==1.11.2
django-easy-audit==1.3.0
django-filter==2.4.0
django-simple-history==3.0.0
django-user-visit==0.4.1
mysqlclient==2.0.3
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
ua-parser==0.10.0
user-agents==2.2.0
